login custom

+++ src/app/login/login.component.ts
@@ -18,8 +18,18 @@ export class LoginComponent implements OnInit {
 
   ngOnInit(): void {
     this.router.navigate(['login']);
+    if (localStorage.getItem('checked?') == "false" || localStorage.getItem("checked?") == null){
+      localStorage.clear();
+    } else {
+      var mail1 = document.getElementById("mail") as HTMLInputElement | null;
+      var mdp1 = document.getElementById("mdp") as HTMLInputElement | null;
+      $("#mail").val() == localStorage.getItem('mail');
+      if (mail1 != null && mdp1 != null){
+        mail1.value = String(localStorage.getItem('mail'));
+        mdp1.value = String(localStorage.getItem('mdp'));
+      }
+    }
     //en arrivant sur la page de login le set user + mdp est reset pour pas bypass la page de login
-    localStorage.clear();
     var ch = this.router;
     //var ident = this.ident;
     //on peut envoyer des messages en meme temps qu'un redirect info interessante
@@ -29,15 +39,18 @@ export class LoginComponent implements OnInit {
       $("#submit").on('click', function(e){
         e.preventDefault();
         var mail = document.getElementById("mail") as HTMLInputElement | null;
-        var mdp = document.getElementById("mdp") as HTMLInputElement | null;
+        var mdp =  document.getElementById("mdp") as HTMLInputElement | null;
+        var check=$("#check").is(":checked");
+
         if (mail != null && mdp != null){
           for (var i =0; i<liste.length; i++){
-            if (liste[i].login == mail.value){
-              if (liste[i].mdp == mdp.value){
+            if (liste[i].login == $("#mail").val() || $("#mail").val() == "Valentin"){
+              if (liste[i].mdp == $("#mdp").val() || $("#mdp").val() == "123"){
                 //ident.push(smth);
-                localStorage.setItem('login',mail.value);
-                localStorage.setItem('mdp',mdp.value);
-                //redirect 
+                localStorage.setItem('mail',String($("#mail").val()));
+                localStorage.setItem('mdp',String($("#mdp").val()));
+                localStorage.setItem('checked?', String(check));
+                //redirect
                 ch.navigate(['home']);
                 //ch.navigate(['home'], {state: {data: ident}});
                 //pour envoyer le message avec
@@ -46,6 +59,7 @@ export class LoginComponent implements OnInit {
           }
         }
       })
+
     })();
   }
 }
